FORGEPLUGINS-126 Send only resource Ids for associations
Due to the changes in the REST plugin introduced by FORGE-1060,
it is preferable to send only the Ids of the resources instead of
sending the entire top-level resource representation to the server.
This commit address that by ensuring that the AngularJS model
contains only the resource Ids.

The commit also ensures that form dirty state handling is done by
storing only the associated Ids in the AngularJS models.